requirements.txt
aniso8601==9.0.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
docker==6.0.0
Flask==2.2.2
Flask-RESTful==0.3.9
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
psutil==5.9.2
pycryptodome==3.15.0
pyftpdlib==1.5.7
pyparsing==3.0.9
pytz==2022.2.1
pywin32==304
requests==2.28.1
six==1.16.0
urllib3==1.26.12
websocket-client==1.4.1
Werkzeug==2.2.2
